+++ mediation/InMobi/source/plugin/Assets/GoogleMobileAds/Mediation/InMobi/Api/InMobi/InMobi.cs
@@ -15,28 +15,36 @@
 using UnityEngine;
 using System.Collections.Generic;
 
-using GoogleMobileAds.Common.Mediation.InMobi;
-using GoogleMobileAds.Mediation;
+using GoogleMobileAds.Mediation.InMobi;
+using GoogleMobileAds.Mediation.InMobi.Common;
 
-namespace GoogleMobileAds.Api.Mediation.InMobi
+namespace GoogleMobileAds.Mediation.InMobi.Api
 {
     public class InMobi
     {
-        public static readonly IInMobiClient client = GetInMobiClient();
-
-        private static IInMobiClient GetInMobiClient()
-        {
-            return InMobiClientFactory.InMobiInstance();
-        }
+        internal static readonly IInMobiClient client = InMobiClientFactory.InMobiInstance();
 
         public static void UpdateGDPRConsent(Dictionary<string, string> consentObject)
         {
-            if (consentObject == null) {
+            if (consentObject == null)
+            {
                 MonoBehaviour.print("Error: Consent object is null.");
                 return;
             }
 
             client.UpdateGDPRConsent(consentObject);
+        }
+    }
+}
+
+namespace GoogleMobileAds.Api.Mediation.InMobi
+{
+    [System.Obsolete("Use `GoogleMobileAds.Mediation.InMobi.Api.InMobi` instead.")]
+    public class InMobi
+    {
+        public static void UpdateGDPRConsent(Dictionary<string, string> consentObject)
+        {
+            GoogleMobileAds.Mediation.InMobi.Api.InMobi.UpdateGDPRConsent(consentObject);
         }
     }
 }